♻️ refactor: group shared utils into record, date, and routing
- Moved individual utility functions into categorized files
- Created separate files for record, date, and routing utilities
