+++ src/app/router.jsx
@@ -4,8 +4,8 @@ import {
   Route,
   Navigate,
 } from 'react-router-dom';
-import getYearMonth from '../shared/utils/getYearMonth';
-import buildUrlWithPage from '../shared/utils/buildUrlWithPage';
+import { getYearMonth } from '../shared/utils/date';
+import { buildUrlWithPage } from '../shared/utils/routing';
 
 import HomePage from '../pages/HomePage';
 import CalendarPage from '../pages/CalendarPage';
