+++ src/pages/HomePage.jsx
@@ -1,9 +1,8 @@
 import { useReducer, useMemo, useState } from 'react';
 import { useOutletContext } from 'react-router-dom';
 import useCurrentYearMonthNumber from '../shared/hooks/useCurrentYearMonthNumber';
-import { isSameYearMonth } from '../shared/utils/isSameYearMonth';
-import deepEqual from '../shared/utils/deepEqual';
-import getTimestamp from '../shared/utils/getTimestamp';
+import { isSameYearMonth, getTimestamp } from '../shared/utils/date';
+import { deepEqual } from '../shared/utils/record';
 import { v4 as uuidv4 } from 'uuid';
 import initialFormState from '../features/form/reducers/initialFormState';
 import formReducer from '../features/form/reducers/formReducer';
