--- src/shared/utils/getActiveTabFromPath.js
@@ -1,3 +1,0 @@
-export function getActiveTabFromPath(pathname) {
-  return pathname.split('/').filter(Boolean)[0];
-}
