--- src/shared/utils/getTimestamp.js
@@ -1,4 +1,0 @@
-// example : "2025-04-22T13:00:00Z"
-export default function getTimestamp(date) {
-  return date.toISOString();
-}
