--- src/shared/utils/getYearMonth.js
@@ -1,7 +1,0 @@
-const getYearMonth = (date) => {
-  const year = date.getFullYear();
-  const month = date.getMonth() + 1;
-  return { year, month };
-};
-
-export default getYearMonth;
