--- src/shared/utils/isSameYearMonth.js
@@ -1,6 +1,0 @@
-export function isSameYearMonth(currentYear, currentMonth, targetDate) {
-  const date = new Date(targetDate);
-  return (
-    date.getFullYear() === currentYear && date.getMonth() + 1 === currentMonth
-  );
-}
