+++ src/shared/utils/record.js
@@ -1,4 +1,18 @@
-export default function deepEqual(a, b) {
+export function getSortedRecordsByDate(records) {
+  return [...records].sort((a, b) => {
+    // 1. date 비교 (문자열: YYYY-MM-DD)
+    if (a.date > b.date) return -1;
+    if (a.date < b.date) return 1;
+
+    // 2. date가 같을 경우, createdAt 비교 (ISO 문자열)
+    if (a.createdAt > b.createdAt) return -1;
+    if (a.createdAt < b.createdAt) return 1;
+
+    return 0; // 둘 다 같으면 순서 유지
+  });
+}
+
+export function deepEqual(a, b) {
   if (a === b) return true;
 
   if (
@@ -29,5 +43,5 @@ export default function deepEqual(a, b) {
 
   ⚠️ 향후 중첩 객체, 배열, 특수 타입(Date, NaN 등) 지원 필요 시 아래처럼 교체
   // import equal from 'fast-deep-equal';
-  // export default equal;
+  // export  equal;
 */
